+++ packages/section/__tests__/unit/utils.test.js
@@ -1,5 +1,9 @@
 import { mockPuzzleSlice } from "@times-components/fixture-generator";
-import { splitPuzzlesBySlices, buildSliceData } from "../../src/utils";
+import {
+  splitPuzzlesBySlices,
+  buildSliceData,
+  getImage
+} from "../../src/utils";
 
 describe("splitPuzzlesBySlices", () => {
   it("should split 3 puzzles in only one slice", () => {
@@ -49,5 +53,32 @@ describe("buildSliceData", () => {
 
     expect(withElementId[0].name).toBe(slices[0].name);
     expect(withElementId[1].name).toBe(slices[1].name);
+  });
+});
+
+describe("getImage", () => {
+  it("should return empty object when cover.crops is empty array", () => {
+    expect(getImage({ crops: [] })).toEqual({});
+  });
+  it("should return empty object when cover.crops is undefined", () => {
+    expect(getImage({})).toEqual({});
+  });
+  it("should return image and ratio", () => {
+    const exampleUrl =
+      "https://www.thetimes.co.uk/imageserver/image/%2Fmethode%2Fsundaytimes%2Fprod%2Fweb%2Fbin%2F9de951a0-5249-11e9-82c1-b5b0bbeb9bfd.jpg?crop=1050%2C1290%2C0%2C0";
+    const cover = {
+      crops: [
+        {
+          ratio: "35:43",
+          url: exampleUrl
+        }
+      ],
+      id: "dummy-cover-id"
+    };
+
+    expect(getImage(cover)).toEqual({
+      ratio: 35 / 43,
+      url: exampleUrl
+    });
   });
 });