+++ packages/article-comments/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.4.56](https://github.com/newsuk/times-components/compare/@times-components/article-comments@0.4.55...@times-components/article-comments@0.4.56) (2019-04-12)
+
+**Note:** Version bump only for package @times-components/article-comments
+
+
+
+
+
 ## [0.4.55](https://github.com/newsuk/times-components/compare/@times-components/article-comments@0.4.54...@times-components/article-comments@0.4.55) (2019-04-12)
 
 **Note:** Version bump only for package @times-components/article-comments