+++ packages/interactive-wrapper/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.2.35](https://github.com/newsuk/times-components/compare/@times-components/interactive-wrapper@0.2.34...@times-components/interactive-wrapper@0.2.35) (2019-04-12)
+
+**Note:** Version bump only for package @times-components/interactive-wrapper
+
+
+
+
+
 ## [0.2.34](https://github.com/newsuk/times-components/compare/@times-components/interactive-wrapper@0.2.33...@times-components/interactive-wrapper@0.2.34) (2019-04-12)
 
 **Note:** Version bump only for package @times-components/interactive-wrapper