+++ packages/provider-test-tools/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [1.16.3](https://github.com/newsuk/times-components/compare/@times-components/provider-test-tools@1.16.2...@times-components/provider-test-tools@1.16.3) (2019-04-12)
+
+**Note:** Version bump only for package @times-components/provider-test-tools
+
+
+
+
+
 ## [1.16.2](https://github.com/newsuk/times-components/compare/@times-components/provider-test-tools@1.16.1...@times-components/provider-test-tools@1.16.2) (2019-04-12)